requirements.txt
cryptography==3.4.8
Django==3.2.8
django-cors-headers==3.10.0
PyJWT==2.3.0
bcrypt==3.2.0
requests==2.26.0